Fix decompose type aliases, stack overflow during Marshal. (#53)
This patch adds missing cases to the decompose reflectValue method to
handle conversion of type aliases to their underlying type. This fixes
behavior in alt.Decompose and also prevents a stack overflow occurring
with `oj.JSON` and `oj.Marshal` where an infinite loop occurs trying to
decompose a type alias to a Go builtin type.